erFindContacts: Requests *all* contacts of address books
Before, we requested only one range of available contacts (the range size was determined by the number of contacts possible to send in one EWS request).

Now, we do multiple requests up to reaching all contacts of the address book (using small batches instead of big ones). For now, the number of contacts for each request is hardcoded to 300, maybe we should allow to configure it.